+++ rllib/core/rl_trainer/scaling_config.py
@@ -11,9 +11,6 @@ if TYPE_CHECKING:
     from ray.rllib.algorithms.algorithm import AlgorithmConfig
 
 
-HyperparamType = Union["AlgorithmConfig", Hyperparams]
-
-
 @dataclass
 class RLModuleBackendConfig:
     """Base class for scaling config relevant to RLTrainer.
@@ -116,63 +113,3 @@ class TrainerScalingConfig:
     num_workers: int = 0
     num_cpus_per_worker: int = 1
     num_gpus_per_worker: int = 0
-
-
-@dataclass
-class RLTrainerSpec:
-    """The spec for construcitng RLTrainer actors.
-
-    Args:
-        rl_trainer_class: The RLTrainer class to use.
-        module_spec: The underlying (MA)RLModule spec to completely define the module.
-        module: Alternatively the RLModule instance can be passed in directly. This
-            only works if the RLTrainer is not an actor.
-        backend_config: The backend config for properly distributing the RLModule.
-        optimizer_config: The optimizer setting to apply during training.
-        trainer_hyperparameters: The extra config for the loss/additional update. The
-            items within this object should be accessible via a dot notation. For
-            example, if the trainer_hyperparameters contains {"coeff": 0.001}, then the
-            learning rate can be accessed via trainer_hyperparameters.coeff. This is
-            useful for passing in algorithm config or a HyperParams that contains the
-            hyper-parameters.
-    """
-
-    rl_trainer_class: Type["RLTrainer"]
-    module_spec: Union["SingleAgentRLModuleSpec", "MultiAgentRLModuleSpec"] = None
-    module: Optional["RLModule"] = None
-    module_backend_config: "RLModuleBackendConfig" = None
-    optimizer_config: Dict[str, Any] = field(default_factory=dict)
-    trainer_hyperparameters: HyperparamType = field(default_factory=dict)
-
-    def __post_init__(self):
-        # convert to hyper params object if needed
-        if isinstance(self.trainer_hyperparameters, dict):
-            self.trainer_hyperparameters = Hyperparams(self.trainer_hyperparameters)
-
-        # if module_backend_config is not set, we will create a dafault.
-        if self.module_backend_config is None:
-            if self.module is not None:
-                if isinstance(self.module, TorchRLModule):
-                    self.module_backend_config = TorchRLModuleBackendConfig()
-                else:
-                    self.module_backend_config = TfRLModuleBackendConfig()
-
-            if self.module_spec is not None:
-                if issubclass(self.module_spec.module_class, TorchRLModule):
-                    self.module_backend_config = TorchRLModuleBackendConfig()
-                else:
-                    self.module_backend_config = TfRLModuleBackendConfig()
-
-    def get_params_dict(self) -> Dict[str, Any]:
-        """Returns the parameters than be passed to the RLTrainer constructor."""
-        return {
-            "module": self.module,
-            "module_spec": self.module_spec,
-            "module_backend_config": self.module_backend_config,
-            "optimizer_config": self.optimizer_config,
-            "trainer_hyperparameters": self.trainer_hyperparameters,
-        }
-
-    def build(self) -> "RLTrainer":
-        """Builds the RLTrainer instance."""
-        return self.rl_trainer_class(**self.get_params_dict())
