+++ rllib/core/rl_trainer/tests/test_trainer_runner.py
@@ -5,7 +5,7 @@ import time
 
 from ray.rllib.policy.sample_batch import DEFAULT_POLICY_ID, MultiAgentBatch
 from ray.rllib.utils.test_utils import get_cartpole_dataset_reader
-from ray.rllib.core.rl_trainer.rl_trainer_config import TrainerScalingConfig
+from ray.rllib.core.rl_trainer.scaling_config import TrainerScalingConfig
 from ray.rllib.core.testing.utils import (
     get_trainer_runner,
     add_module_to_runner_or_trainer,
